migration changes

+++ src/database/migrations/2026_02_12_150130_add_new_column_object_key_in_files_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
      */
     public function down(): void
     {
-        if (Schema::hasColumn('files', 'object_key')) {
+        if ( (Schema::hasTable('files')) && (Schema::hasColumn('files', 'object_key'))) {
             Schema::table('files', function (Blueprint $table) {
                 $table->dropColumn(['object_key', 'is_entity']);
             });
